+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.15.4 (TBD)
+
+- Implemented `LexicographicWord` struct ([#443](https://github.com/0xMiden/crypto/pull/443/)).
+- Added `SequentialCommit` trait ([#443](https://github.com/0xMiden/crypto/pull/443/)).
+
 ## 0.15.3 (2025-06-18)
 
 - Fixed conversion error from a slice of bytes into `Word`.